o/configcore: fix nomanagers test
Fix nomanagers test by making sure tests for config groups that are
not compiled for nomanagers are not not run in that case and by adding
additional mocking for bits not compiled for nomanagers but that were
used in tests that need to run for nomanagers case.
