+++ overlord/configstate/configcore/corecfg_test.go
@@ -175,18 +175,6 @@ type runCfgSuite struct {
 
 var _ = Suite(&runCfgSuite{})
 
-func (r *runCfgSuite) TestConfigureUnknownOption(c *C) {
-	conf := &mockConf{
-		state: r.state,
-		changes: map[string]interface{}{
-			"unknown.option": "1",
-		},
-	}
-
-	err := configcore.Run(coreDev, conf)
-	c.Check(err, ErrorMatches, `cannot set "core.unknown.option": unsupported system option`)
-}
-
 type mockDev struct {
 	mode    string
 	classic bool
@@ -262,10 +250,7 @@ func (s *applyCfgSuite) TestPlainCoreConfigGetMaybe(c *C) {
 	c.Check(val, DeepEquals, "bar")
 }
 
-func (s *applyCfgSuite) TestNilHandlePanics(c *C) {
+func (s *applyCfgSuite) TestNilHandleAddFSOnlyHandlerPanic(c *C) {
 	c.Assert(func() { configcore.AddFSOnlyHandler(nil, nil, nil) },
 		Panics, "cannot have nil handle with fsOnlyHandler")
-
-	c.Assert(func() { configcore.AddWithStateHandler(nil, nil, nil) },
-		Panics, "cannot have nil handle with addWithStateHandler if validatedOnlyStateConfig flag is not set")
 }
