+++ overlord/configstate/configcore/experimental_test.go
@@ -47,7 +47,7 @@ func (s *experimentalSuite) TestConfigureExperimentalSettingsInvalid(c *C) {
 			state:   s.state,
 			changes: map[string]interface{}{featureConf(feature): "foo"},
 		}
-		err := configcore.Run(classicDev, conf)
+		err := configcore.MockRun(classicDev, conf)
 		c.Check(err, ErrorMatches, fmt.Sprintf(`%s can only be set to 'true' or 'false'`, featureConf(feature)))
 	}
 }
@@ -59,7 +59,7 @@ func (s *experimentalSuite) TestConfigureExperimentalSettingsHappy(c *C) {
 				state: s.state,
 				conf:  map[string]interface{}{featureConf(feature): t},
 			}
-			err := configcore.Run(classicDev, conf)
+			err := configcore.MockRun(classicDev, conf)
 			c.Check(err, IsNil)
 		}
 	}
@@ -70,12 +70,12 @@ func (s *experimentalSuite) TestExportedFeatures(c *C) {
 		state: s.state,
 		conf:  map[string]interface{}{featureConf(features.PerUserMountNamespace): true},
 	}
-	err := configcore.Run(classicDev, conf)
+	err := configcore.MockRun(classicDev, conf)
 	c.Assert(err, IsNil)
 	c.Check(features.PerUserMountNamespace.ControlFile(), testutil.FilePresent)
 
 	delete(conf.changes, "experimental.per-user-mount-namespace")
-	err = configcore.Run(classicDev, conf)
+	err = configcore.MockRun(classicDev, conf)
 	c.Assert(err, IsNil)
 	c.Check(features.PerUserMountNamespace.ControlFile(), testutil.FilePresent)
 }
