+++ overlord/configstate/configcore/powerbtn_test.go
@@ -54,7 +54,7 @@ func (s *powerbtnSuite) TestConfigurePowerButtonInvalid(c *C) {
 func (s *powerbtnSuite) TestConfigurePowerIntegration(c *C) {
 	for _, action := range []string{"ignore", "poweroff", "reboot", "halt", "kexec", "suspend", "hibernate", "hybrid-sleep", "lock"} {
 
-		err := configcore.Run(coreDev, &mockConf{
+		err := configcore.MockRun(coreDev, &mockConf{
 			state: s.state,
 			conf: map[string]interface{}{
 				"system.power-key-action": action,
