+++ overlord/configstate/configcore/refresh_test.go
@@ -1,4 +1,6 @@
 // -*- Mode: Go; indent-tabs-mode: t -*-
+//go:build !nomanagers
+// +build !nomanagers
 
 /*
  * Copyright (C) 2017-2018 Canonical Ltd
