+++ overlord/configstate/configcore/swap_test.go
@@ -55,7 +55,7 @@ func (s *swapCfgSuite) SetUpTest(c *C) {
 
 func (s *swapCfgSuite) TestConfigureSwapSizeOnlyWhenChanged(c *C) {
 	// set it to 1M initially
-	err := configcore.Run(coreDev, &mockConf{
+	err := configcore.MockRun(coreDev, &mockConf{
 		state: s.state,
 		changes: map[string]interface{}{
 			"swap.size": "1048576",
@@ -76,7 +76,7 @@ SIZE=1
 	s.systemctlArgs = nil
 
 	// running it with the same changes as conf results in no calls to systemd
-	err = configcore.Run(coreDev, &mockConf{
+	err = configcore.MockRun(coreDev, &mockConf{
 		state: s.state,
 		conf: map[string]interface{}{
 			"swap.size": "1048576",
@@ -96,7 +96,7 @@ SIZE=1
 
 func (s *swapCfgSuite) TestConfigureSwapSize(c *C) {
 	// set it to 1M initially
-	err := configcore.Run(coreDev, &mockConf{
+	err := configcore.MockRun(coreDev, &mockConf{
 		state: s.state,
 		changes: map[string]interface{}{
 			"swap.size": "1048576",
@@ -117,7 +117,7 @@ SIZE=1
 	s.systemctlArgs = nil
 
 	// now change it to empty
-	err = configcore.Run(coreDev, &mockConf{
+	err = configcore.MockRun(coreDev, &mockConf{
 		state: s.state,
 		conf: map[string]interface{}{
 			"swap.size": "1048576",
