+++ overlord/configstate/configcore/tmp_test.go
@@ -57,8 +57,7 @@ func (s *tmpfsSuite) TestConfigureTmpfsGoodVals(c *C) {
 	defer mountCmd.Restore()
 
 	for _, size := range []string{"104857600", "16M", "7G", "0"} {
-
-		err := configcore.Run(coreDev, &mockConf{
+		err := configcore.MockRun(coreDev, &mockConf{
 			state: s.state,
 			conf: map[string]interface{}{
 				"tmp.size": size,
@@ -81,7 +80,7 @@ func (s *tmpfsSuite) TestConfigureTmpfsBadVals(c *C) {
 	for _, size := range []string{"100p", "0x123", "10485f7600", "20%%",
 		"20%", "100m", "10k", "10K", "10g"} {
 
-		err := configcore.Run(coreDev, &mockConf{
+		err := configcore.MockRun(coreDev, &mockConf{
 			state: s.state,
 			conf: map[string]interface{}{
 				"tmp.size": size,
@@ -99,7 +98,7 @@ func (s *tmpfsSuite) TestConfigureTmpfsBadVals(c *C) {
 func (s *tmpfsSuite) TestConfigureTmpfsTooSmall(c *C) {
 	for _, size := range []string{"1", "16777215"} {
 
-		err := configcore.Run(coreDev, &mockConf{
+		err := configcore.MockRun(coreDev, &mockConf{
 			state: s.state,
 			conf: map[string]interface{}{
 				"tmp.size": size,
@@ -124,7 +123,7 @@ func (s *tmpfsSuite) TestConfigureTmpfsAllConfDirExistsAlready(c *C) {
 	c.Assert(err, IsNil)
 
 	size := "100M"
-	err = configcore.Run(coreDev, &mockConf{
+	err = configcore.MockRun(coreDev, &mockConf{
 		state: s.state,
 		conf: map[string]interface{}{
 			"tmp.size": size,
@@ -156,7 +155,7 @@ func (s *tmpfsSuite) TestConfigureTmpfsNoFileUpdate(c *C) {
 	// To make sure the times will differ if the file is newly written
 	time.Sleep(100 * time.Millisecond)
 
-	err = configcore.Run(coreDev, &mockConf{
+	err = configcore.MockRun(coreDev, &mockConf{
 		state: s.state,
 		conf: map[string]interface{}{
 			"tmp.size": size,
@@ -189,7 +188,7 @@ func (s *tmpfsSuite) TestConfigureTmpfsRemovesIfUnset(c *C) {
 	err = ioutil.WriteFile(s.servOverridePath, []byte(content), 0644)
 	c.Assert(err, IsNil)
 
-	err = configcore.Run(coreDev, &mockConf{
+	err = configcore.MockRun(coreDev, &mockConf{
 		state: s.state,
 		conf: map[string]interface{}{
 			"tmp.size": "",
