+++ overlord/configstate/configcore/watchdog_test.go
@@ -51,7 +51,7 @@ func (s *watchdogSuite) SetUpTest(c *C) {
 func (s *watchdogSuite) TestConfigureWatchdog(c *C) {
 	for option, val := range map[string]string{"runtime-timeout": "10", "shutdown-timeout": "60"} {
 
-		err := configcore.Run(coreDev, &mockConf{
+		err := configcore.MockRun(coreDev, &mockConf{
 			state: s.state,
 			conf: map[string]interface{}{
 				fmt.Sprintf("watchdog.%s", option): val + "s",
@@ -84,7 +84,7 @@ func (s *watchdogSuite) TestConfigureWatchdogUnits(c *C) {
 	}
 
 	for _, tunit := range []timeUnit{{"s", 1}, {"m", 60}, {"h", 3600}} {
-		err := configcore.Run(coreDev, &mockConf{
+		err := configcore.MockRun(coreDev, &mockConf{
 			state: s.state,
 			conf: map[string]interface{}{
 				"watchdog.runtime-timeout":  fmt.Sprintf("%d", times[0]) + tunit.unit,
@@ -100,7 +100,7 @@ func (s *watchdogSuite) TestConfigureWatchdogUnits(c *C) {
 
 func (s *watchdogSuite) TestConfigureWatchdogAll(c *C) {
 	times := []int{10, 100}
-	err := configcore.Run(coreDev, &mockConf{
+	err := configcore.MockRun(coreDev, &mockConf{
 		state: s.state,
 		conf: map[string]interface{}{
 			"watchdog.runtime-timeout":  fmt.Sprintf("%ds", times[0]),
@@ -123,7 +123,7 @@ func (s *watchdogSuite) TestConfigureWatchdogAllConfDirExistsAlready(c *C) {
 	c.Assert(err, IsNil)
 
 	times := []int{10, 100}
-	err = configcore.Run(coreDev, &mockConf{
+	err = configcore.MockRun(coreDev, &mockConf{
 		state: s.state,
 		conf: map[string]interface{}{
 			"watchdog.runtime-timeout":  fmt.Sprintf("%ds", times[0]),
@@ -148,7 +148,7 @@ func (s *watchdogSuite) TestConfigureWatchdogBadFormat(c *C) {
 	for _, badVal := range []badValErr{{"BAD", ".*invalid duration.*"},
 		{"-5s", ".*negative duration.*"},
 		{"34k", ".*unknown unit.*"}} {
-		err := configcore.Run(coreDev, &mockConf{
+		err := configcore.MockRun(coreDev, &mockConf{
 			state: s.state,
 			conf: map[string]interface{}{
 				"watchdog.runtime-timeout": badVal.val,
@@ -178,7 +178,7 @@ func (s *watchdogSuite) TestConfigureWatchdogNoFileUpdate(c *C) {
 	// To make sure the times will defer if the file is newly written
 	time.Sleep(100 * time.Millisecond)
 
-	err = configcore.Run(coreDev, &mockConf{
+	err = configcore.MockRun(coreDev, &mockConf{
 		state: s.state,
 		conf: map[string]interface{}{
 			"watchdog.runtime-timeout":  fmt.Sprintf("%ds", times[0]),
@@ -210,7 +210,7 @@ ShutdownWatchdogSec=20
 	err = ioutil.WriteFile(s.mockEtcEnvironment, []byte(content), 0644)
 	c.Assert(err, IsNil)
 
-	err = configcore.Run(coreDev, &mockConf{
+	err = configcore.MockRun(coreDev, &mockConf{
 		state: s.state,
 		conf: map[string]interface{}{
 			"watchdog.runtime-timeout":  0,
